src/main/java/domain/Radio.java
package domain;

public class Radio {
    private int maxStation = 10;
    private int minStation;
    private int currentStation;
    private int maxVolume = 100;
    private int minVolume;
    private int currentVolume;


    public void setCurrentStation(int newCurrentStation) {
        if (newCurrentStation < minStation) {
            return;
        }
        if (newCurrentStation > maxStation) {
            return;
        }
        this.currentStation = newCurrentStation;
    }

    public int getCurrentStation() {
        return currentStation;
    }

    public void increaseStation() {
        if (currentStation <= 9) {
            currentStation = currentStation + 1;
        }
        this.currentStation = currentStation;
    }

    public void next() {
        if (currentStation >= 9) {
            currentStation = 0;
        }
        this.currentStation = currentStation;
    }

    public void prev() {
        if (currentStation <= 0) {
            currentStation = 9;
        }
        this.currentStation = currentStation;
    }

    public Radio(int amount) {

        this.minStation = minStation;
        this.maxStation = maxStation;

    }

    public Radio() {

        this.maxStation = 10;
    }

    public void setCurrentVolume(int newCurrentVolume) {
        if (newCurrentVolume < minVolume) {
            return;
        }
        if (newCurrentVolume > maxVolume) {
            return;
        }
        this.currentVolume = newCurrentVolume;
    }

    public int getCurrentVolume() {
        return currentVolume;
    }

    public void increaseVolume() {
        if (currentVolume <= 100) {
            currentVolume = currentVolume + 1;
        }
        setCurrentVolume(currentVolume);
    }

    public void maxVolume() {
        if (currentVolume >= 100) {
            currentVolume = 100;
        }
        this.currentVolume = currentVolume;
    }

    public void decreaseVolume() {
        if (currentVolume <= 0) {
            currentVolume = 0;
        }
        this.currentVolume = currentVolume;
    }


    public int getMinStation() {
        return minStation;
    }

    public void setMinStation(int minStation) {
        this.minStation = minStation;
    }

    public int getMaxStation() {
        return maxStation;
    }

    public void setMaxStation(int maxStation) {
        this.maxStation = maxStation;
    }


    public int getMaxVolume() {
        return maxVolume;
    }

    public void setMaxVolume(int maxVolume) {
        this.maxVolume = maxVolume;
    }

    public int getMinVolume() {
        return minVolume;
    }

    public void setMinVolume(int minVolume) {
        this.minVolume = minVolume;
    }

}


